refactor: Typescript
