+++ webpack.config.js
@@ -1,37 +1,1 @@
-const path = require('path');
-const webpack = require('webpack');
-
-module.exports = {
-  entry: './src/siema.js',
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: 'siema.min.js',
-    library: 'Siema',
-    libraryTarget: 'umd',
-    umdNamedDefine: true
-  },
-  module: {
-    rules: [
-      {
-        enforce: 'pre',
-        test: /\.(js|jsx)$/,
-        exclude: /node_modules/,
-        loader: 'eslint-loader',
-      },
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /(node_modules)/,
-        loader: 'babel-loader',
-        query: {
-          presets: ['es2015'],
-          plugins: ['babel-plugin-add-module-exports'],
-        },
-      },
-    ]
-  },
-  plugins: [
-    new webpack.optimize.UglifyJsPlugin({
-      minimize: true,
-    }),
-  ],
-};
+module.exports = require('./config/webpack.config.js');
